refactor: ProductContainer 내부 UI 컴포넌트 분리
