--- src/refactoring/components/admin/AdminProductList.tsx
@@ -1,148 +1,0 @@
-import { Product } from "../../../types";
-import { DiscountInput } from "./DiscountInput";
-
-interface IProps {
-  products: Product[];
-  openProductIds: Set<string>;
-  editingProduct: Product | null;
-  newDiscount: { quantity: number; rate: number };
-  toggleProductAccordion: (id: string) => void;
-  handleEditProduct: (product: Product) => void;
-  handleProductNameUpdate: (id: string, name: string) => void;
-  handlePriceUpdate: (id: string, price: number) => void;
-  handleStockUpdate: (id: string, stock: number) => void;
-  handleAddDiscount: (id: string) => void;
-  handleRemoveDiscount: (id: string, index: number) => void;
-  handleEditComplete: () => void;
-  setNewDiscount: (discount: { quantity: number; rate: number }) => void;
-}
-
-const AdminProductList = ({
-  products,
-  openProductIds,
-  editingProduct,
-  newDiscount,
-  toggleProductAccordion,
-  handleEditProduct,
-  handleProductNameUpdate,
-  handlePriceUpdate,
-  handleStockUpdate,
-  handleAddDiscount,
-  handleRemoveDiscount,
-  handleEditComplete,
-  setNewDiscount,
-}: IProps) => {
-  return (
-    <div className="space-y-2">
-      {products.map((product, index) => (
-        <div
-          key={product.id}
-          data-testid={`product-${index + 1}`}
-          className="bg-white p-4 rounded shadow"
-        >
-          <button
-            data-testid="toggle-button"
-            onClick={() => toggleProductAccordion(product.id)}
-            className="w-full text-left font-semibold"
-          >
-            {product.name} - {product.price}원 (재고: {product.stock})
-          </button>
-          {openProductIds.has(product.id) && (
-            <div className="mt-2">
-              {editingProduct && editingProduct.id === product.id ? (
-                <div>
-                  <div className="mb-4">
-                    <label className="block mb-1">상품명: </label>
-                    <input
-                      type="text"
-                      value={editingProduct.name}
-                      onChange={(e) =>
-                        handleProductNameUpdate(product.id, e.target.value)
-                      }
-                      className="w-full p-2 border rounded"
-                    />
-                  </div>
-                  <div className="mb-4">
-                    <label className="block mb-1">가격: </label>
-                    <input
-                      type="number"
-                      value={editingProduct.price}
-                      onChange={(e) =>
-                        handlePriceUpdate(product.id, parseInt(e.target.value))
-                      }
-                      className="w-full p-2 border rounded"
-                    />
-                  </div>
-                  <div className="mb-4">
-                    <label className="block mb-1">재고: </label>
-                    <input
-                      type="number"
-                      value={editingProduct.stock}
-                      onChange={(e) =>
-                        handleStockUpdate(product.id, parseInt(e.target.value))
-                      }
-                      className="w-full p-2 border rounded"
-                    />
-                  </div>
-                  <div>
-                    <h4 className="text-lg font-semibold mb-2">할인 정보</h4>
-                    {editingProduct.discounts.map((discount, index) => (
-                      <div
-                        key={index}
-                        className="flex justify-between items-center mb-2"
-                      >
-                        <span>
-                          {discount.quantity}개 이상 구매 시{" "}
-                          {discount.rate * 100}% 할인
-                        </span>
-                        <button
-                          onClick={() =>
-                            handleRemoveDiscount(product.id, index)
-                          }
-                          className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
-                        >
-                          삭제
-                        </button>
-                      </div>
-                    ))}
-                    <DiscountInput
-                      newDiscount={newDiscount}
-                      onChange={setNewDiscount}
-                      onAdd={() => handleAddDiscount(product.id)}
-                    />
-                  </div>
-                  <button
-                    onClick={handleEditComplete}
-                    className="bg-green-500 text-white px-2 py-1 rounded hover:bg-green-600 mt-2"
-                  >
-                    수정 완료
-                  </button>
-                </div>
-              ) : (
-                <div>
-                  {product.discounts.map((discount, index) => (
-                    <div key={index} className="mb-2">
-                      <span>
-                        {discount.quantity}개 이상 구매 시 {discount.rate * 100}
-                        % 할인
-                      </span>
-                    </div>
-                  ))}
-                  <button
-                    data-testid="modify-button"
-                    onClick={() => handleEditProduct(product)}
-                    className="bg-blue-500 text-white px-2 py-1 rounded hover:bg-blue-600 mt-2"
-                  >
-                    수정
-                  </button>
-                </div>
-              )}
-            </div>
-          )}
-        </div>
-      ))}
-    </div>
-  );
-};
-
-export default AdminProductList;
