+++ src/refactoring/components/admin/index.ts
@@ -1,2 +1,4 @@
 export { default as NewProductSection } from "./NewProductSection";
-export { default as AdminProductList } from "./AdminProductList";
+export { default as ProductContainer } from "./ProductContainer";
+export { default as CouponManager } from "./CouponManager";
+export { default as ProductEditor } from "./ProductEditor";
