+++ src/refactoring/components/coupon/index.ts
@@ -1,4 +1,3 @@
 export { default as CouponContainer } from "./CouponContainer";
 export { default as CouponSelector } from "./CouponSelector";
 export { default as AppliedCouponInfo } from "./AppliedCouponInfo";
-export { default as CouponManager } from "./CouponManager";
